+++ apps/etterlatte-behandling/src/main/kotlin/saksbehandler/SaksbehandlerService.kt
@@ -1,7 +1,10 @@
 package no.nav.etterlatte.saksbehandler
 
+import kotlinx.coroutines.runBlocking
 import no.nav.etterlatte.Kontekst
+import no.nav.etterlatte.behandling.klienter.AxsysKlient
 import no.nav.etterlatte.behandling.klienter.SaksbehandlerInfo
+import no.nav.etterlatte.inTransaction
 
 data class Saksbehandler(
     val ident: String,
@@ -12,22 +15,54 @@ data class Saksbehandler(
     val kanSeOppgaveliste: Boolean,
 )
 
-class SaksbehandlerService(private val dao: SaksbehandlerInfoDao) {
-    fun hentKomplettSaksbehandler(ident: String): Saksbehandler {
+data class SaksbehandlerEnhet(
+    val enhetsNummer: String,
+    val navn: String,
+)
+
+interface SaksbehandlerService {
+    fun hentKomplettSaksbehandler(ident: String): Saksbehandler
+
+    fun hentSaksbehandlereForEnhet(enhet: List<String>): Set<SaksbehandlerInfo>
+
+    fun hentEnheterForSaksbehandlerIdentWrapper(ident: String): List<SaksbehandlerEnhet>
+}
+
+class SaksbehandlerServiceImpl(
+    private val dao: SaksbehandlerInfoDao,
+    private val axsysKlient: AxsysKlient,
+) : SaksbehandlerService {
+    override fun hentKomplettSaksbehandler(ident: String): Saksbehandler {
         val innloggetSaksbehandler = Kontekst.get().appUserAsSaksbehandler()
 
         val saksbehandlerNavn: String? = dao.hentSaksbehandlerNavn(ident)
 
         return Saksbehandler(
-            ident = ident,
-            navn = if (!saksbehandlerNavn.isNullOrEmpty()) saksbehandlerNavn else ident,
-            enheter = innloggetSaksbehandler.enheter(),
-            kanAttestere = innloggetSaksbehandler.saksbehandlerMedRoller.harRolleAttestant(),
+            ident,
+            if (!saksbehandlerNavn.isNullOrEmpty()) saksbehandlerNavn else ident,
+            innloggetSaksbehandler.enheter(),
+            innloggetSaksbehandler.saksbehandlerMedRoller.harRolleAttestant(),
             skriveEnheter = innloggetSaksbehandler.enheterMedSkrivetilgang(),
             kanSeOppgaveliste = innloggetSaksbehandler.kanSeOppgaveBenken(),
         )
     }
 
-    fun hentSaksbehandlereForEnhet(enhet: List<String>): Set<SaksbehandlerInfo> =
+    override fun hentSaksbehandlereForEnhet(enhet: List<String>): Set<SaksbehandlerInfo> =
         enhet.flatMap { dao.hentSaksbehandlereForEnhet(it) }.toSet()
+
+    override fun hentEnheterForSaksbehandlerIdentWrapper(ident: String): List<SaksbehandlerEnhet> {
+        val harIntransaction = Kontekst.get().databasecontxt.harIntransaction()
+        return if (harIntransaction) {
+            hentEnheterForSaksbehandler(ident)
+        } else {
+            inTransaction {
+                hentEnheterForSaksbehandler(ident)
+            }
+        }
+    }
+
+    private fun hentEnheterForSaksbehandler(ident: String): List<SaksbehandlerEnhet> {
+        return dao.hentSaksbehandlerEnheter(ident)
+            ?: runBlocking { axsysKlient.hentEnheterForIdent(ident) }
+    }
 }
